Changed Navigation Bar Organization
