+++ styles/globals.scss
@@ -112,13 +112,6 @@ h4 {
   line-height: 1.04em;
 }
 
-h5,
-.nav-links a {
-  font: {
-    size: 1.8rem;
-    weight: 600;
-  }
-}
 
 .S1 {
   font-size: 1.7rem;
@@ -178,15 +171,6 @@ h5,
     line-height: 1.375em;
   }
 
-  h5,
-  .nav-links a {
-    font: {
-      size: 1.6rem;
-      weight: 600;
-    }
-    line-height: 1.375em;
-  }
-
   .S1 {
     font-size: 1.8rem;
     line-height: 1.5em;
@@ -216,113 +200,6 @@ h5,
   }
 }
 
-/*------NAVIGATION SECTION ------*/
-.navbar {
-  background-color: var(--color-white) !important;
-  box-shadow: rgba(0, 0, 0, 0.15) 1.95px 1.95px 2.6px;
-  display: flex;
-  font-size: 1.125rem;
-  font-weight: 600;
-  height: 125px;
-  top: 0;
-}
-
-a.nav-link {
-  display: flex;
-  margin: 0 auto;
-}
-
-//Navigation - Logo Wrapper
-.navbar-logo-wrapper {
-  margin: 5%;
-}
-
-//Navigation - Links Wrapper
-.navbar-nav-wrapper {
-  margin: 0 auto;
-}
-
-//Navigation - Link Section
-.nav-links {
-  display: inline-flex;
-  width: 100%;
-}
-
-.nav-link {
-  display: inline;
-}
-
-//Navigation - Links
-.nav-links a {
-  display: inline-block;
-  margin: 0 25px;
-}
-
-//Navigation - Programmes Dropdown Links
-a.dropdown-item {
-  color: var(--color-black);
-  font-size: 1rem;
-  font-weight: 500;
-  margin-left: 15px;
-  width: 276px;
-  position: 10px;
-
-  &:hover {
-    width: 90%;
-  }
-}
-
-//Navigation - Donate Button
-.donate-button {
-  background-color: var(--color-honey);
-  border: none;
-  border-radius: 25px;
-  font-weight: 500;
-  font-size: 1rem;
-  height: 46px;
-  margin-left: 20px;
-  padding: 0px 25px;
-  text-transform: capitalize;
-
-  &:hover {
-    background-color: #d0b316;
-    transition: 0.5s;
-  }
-}
-
-// ACTIVE UNDERLINE
-/*.nav-link.active::after {
-  content: "";
-  float: left;
-  background: var(--color-honey);
-  width: 100%;
-  height: 4px;
-  border-radius: 3px;
-  width: 100%;
-} 
-*/
-
-.navbar-expand-lg .navbar-collapse {
-  justify-content: center;
-}
-
-.navbar-light .navbar-nav .nav-link {
-  color: var(--color-black);
-}
-
-@media only screen and (max-width: 992px) {
-  .nav-links {
-    background-color: var(--color-white);
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-  }
-
-  .navbar-nav {
-    justify-content: center;
-  }
-}
 
 /*------HERO SECTION ------*/
 
@@ -776,9 +653,7 @@ span {
   align-items: center;
 }
 
-.navbar-brand {
-  font-size: xx-large;
-}
+
 
 #empower-text {
   color: var(--color-berry);
